requirements.txt
# requirements.txt
discord.py
python-dotenv
PyMySQL